pyproject.toml
[project]
name = "datathon-grupo-03"
version = "0.3.0"
description = "Projeto de risco de crédito"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "numpy>=2.4.3",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
]

[dependency-groups]
eda = [
    "seaborn>=0.13.2",
]

model = [
    "scikit-learn>=1.8.0",
    "torch>=2.11.0",
    "shap>=0.46",
    "numba>=0.60.0",
    "llvmlite>=0.43.0",
]

serving = [
    "fastapi>=0.135.2",
    "uvicorn>=0.42.0",
]

monitoring = [
    "evidently>=0.7.21",
    "prometheus-client>=0.24.1",
    "mlflow>=3.10.1"
]

data = [
    "dvc>=3.67.0",
]

agent = [
    "langchain-core>=0.3",
    "langchain-community>=0.3",
    "langchain-text-splitters>=0.3",
    "langchain-huggingface>=0.1",
    "langchain-groq>=0.3",
    "sentence-transformers>=3.0",
    "chromadb>=0.6",
    "groq>=0.9",
    "python-dotenv>=1.0",
]

experiment = [
    "mlflow>=3.10.1",
]

dev = [
    "black>=26.3.1",
    "flake8>=7.3.0",
    "isort>=8.0.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.1.0",
]

[tool.pytest.ini_options]
pythonpath = [".", "src"]

[tool.isort]
profile = "black"
